+++ src/main/java/org/openhab/binding/mqtt/frigatesvr/internal/helpers/frigateSVRNetworkHelper.java
@@ -16,6 +16,10 @@ import java.net.DatagramSocket;
 import java.net.InetAddress;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
+import org.eclipse.jdt.annotation.Nullable;
+import org.openhab.binding.mqtt.frigatesvr.internal.structures.frigateSVRServices;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * The {@link mqtt.frigateSVRNetworkHelper} provides some useful information
@@ -26,11 +30,29 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 @NonNullByDefault
 public class frigateSVRNetworkHelper {
 
-    private String hostname = "";
-    private String port = "8080";
+    private final Logger logger = LoggerFactory.getLogger(frigateSVRNetworkHelper.class);
+    private frigateSVRServices services;
 
-    public frigateSVRNetworkHelper() {
-        hostname = this.GetAccessName();
+    public frigateSVRNetworkHelper(frigateSVRServices services) {
+        this.services = services;
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    // GetProtocol
+    //
+    // The protocol would be set up in config; we do need to determine if
+    // https is required from the config? Todo.
+
+    public String GetProtocol() {
+        String protocol = "";
+        if (System.getProperty("org.osgi.service.http.port.secure") != null) {
+            protocol = "https";
+        } else {
+            if (System.getProperty("org.osgi.service.http.port") != null) {
+                protocol = "http";
+            }
+        }
+        return protocol;
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -39,7 +61,7 @@ public class frigateSVRNetworkHelper {
     // Returns the host, FQDN or otherwise (no port suffixes etc)
 
     public String GetHost() {
-        return hostname;
+        return this.GetAccessName();
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -48,6 +70,17 @@ public class frigateSVRNetworkHelper {
     // Return the port the current OH instance is running on
 
     public String GetPort() {
+        @Nullable
+        String port;
+        port = System.getProperty("org.osgi.service.http.port.secure");
+        if (port == null) {
+            logger.info("Secure web service port not found, looking for insecure");
+            port = System.getProperty("org.osgi.service.http.port");
+            if (port == null) {
+                logger.error("unable to determine secure and insecure web server ports");
+                port = "";
+            }
+        }
         return port;
     }
 
@@ -57,7 +90,13 @@ public class frigateSVRNetworkHelper {
     // Return the base URL to the running instance
 
     public String GetHostBaseURL() {
-        return "http://" + hostname + ":" + port;
+        String protocol = GetProtocol();
+        String port = GetPort();
+        String url = "";
+        if (!protocol.isBlank() && !port.isBlank()) {
+            url = protocol + "://" + GetHost() + ":" + port;
+        }
+        return url;
     }
 
     ///////////////////////////////////////////////////////////////////////////
@@ -68,26 +107,43 @@ public class frigateSVRNetworkHelper {
     // revert to 'localhost
 
     private String GetAccessName() {
+
         String host = "";
+        logger.info("attempting to obtain host name");
         try {
-            host = InetAddress.getLocalHost().getCanonicalHostName();
+            host = InetAddress.getByName(services.addressService.getPrimaryIpv4HostAddress()).getCanonicalHostName();
         } catch (Exception e) {
             try {
-                host = InetAddress.getLocalHost().getHostAddress();
+                logger.error("trying addressService with getHostAddress");
+                host = InetAddress.getByName(services.addressService.getPrimaryIpv4HostAddress()).getHostAddress();
             } catch (Exception f) {
                 try {
-                    // if all is lost, try this
-                    // the IP does not need to be reachable
-                    DatagramSocket socket = new DatagramSocket();
-                    socket.connect(InetAddress.getByName("8.8.8.8"), 10002);
-                    host = socket.getLocalAddress().getHostAddress();
-                    socket.close();
+                    logger.error("trying getLocalHost with getCanonicalHostName");
+                    host = InetAddress.getLocalHost().getCanonicalHostName();
                 } catch (Exception g) {
-                    // I give up
-                    host = "localhost";
+                    try {
+                        logger.error("failed; trying getLocalHost and getHostAddress");
+                        host = InetAddress.getLocalHost().getHostAddress();
+                    } catch (Exception h) {
+                        try {
+                            logger.error("failed; trying local connection");
+                            // if all is lost, try this
+                            // the IP does not need to be reachable
+                            DatagramSocket socket = new DatagramSocket();
+                            socket.connect(InetAddress.getByName("8.8.8.8"), 10002);
+                            host = socket.getLocalAddress().getHostAddress();
+                            socket.close();
+                        } catch (Exception i) {
+                            // I give up
+                            logger.error(
+                                    "all methods failed; unable to determine forwarder host name - using <unknown>");
+                            host = "<unknown>";
+                        }
+                    }
                 }
             }
         }
+        logger.info("have forwarder hostname {}", host);
         return host;
     }
 }